README.md
This is about building a Doctor Appointment App with using Laravel as Backed and Dashboard. The main idea of this app is allow users to make an appointment with the doctors using mobile app. And, the major features will be demonstrated in the following videos in this series, such as Laravel Dashboard, make appointment with registered doctor, view doctor's details, rating and feedback submission and etc.